requirements.txt
# Core Dependencies
fastapi
uvicorn[standard]
pydantic

# LangChain & AI
langchain
langchain-openai
langchain-community
langgraph
langchain-pinecone

# Vector Store
pinecone

# Document Processing
pypdf
python-multipart

# Utilities
python-dotenv
requests

# Frontend
streamlit